webservice/requirements.txt
# for fastapi
fastapi == 0.115.2
uvicorn == 0.31.1
requests == 2.32.3
python-dotenv == 1.0.1

# for telegram
python-telegram-bot == 20.3
python-telegram-bot[job-queue]

# for AI
openai == 1.51.0
langchain == 0.3.2
langchain-community == 0.3.1
langchain-openai == 0.2.2
tiktoken == 0.8.0
faiss-cpu == 1.8.0.post1
